model-service/requirements.txt
torch
torchvision
pillow==8.1.1
opencv-python
paddlepaddle
paddlehub
waitress
flask
numpy==1.15.2
scikit-image==0.14.0